Respect rect offset in TexturePacker atlases
If a sprite sheet is created using the "Trim" option in TexturePacker,
individual sprites can be offset from the original size. To make them appear
at the right position in an animation, this offset (namely the x and y values
of the spriteColorRect) must be taken into account when setting the deck's
rect.