Add scissorRect support for Group
This allows a group to support setScissorRect() and automatically
apply the rect to child props. It also, if your MOAI tree has
getScissorRect(), avoids returning clipped props in
getTouchableProp. Without that change, a group with a scissor rect
can report clicks on members that are invisible, and if the group is
in front of another prop, it can prevent that prop from receiving
clicks.