Give touch events references to the props that they report being on.
If you touchDown on a prop, then move, then touchUp on another prop, the
first prop still receives the touchUp event, and e.prop is the prop
receiving the event, even though that isn't the prop that the touchUp
occurred on. Add "e.active_prop" and "e.other_prop" fields to allow
UI elements to distinguish these cases.